@@ -1,51 +1,53 @@
-[tool.poetry]
-name = "fastapi-draft"
+[project]
+name = "shop-educational"
 version = "0.1.0"
-description = "template"
-authors = ["alina1412 <alina1412@mail.ru>"]
+description = ""
 readme = "README.md"
-packages = []
+requires-python = ">=3.11"
+dependencies = [
+    "alembic==1.14.0",
+    "annotated-types==0.7.0",
+    "anyio==4.8.0",
+    "astroid==3.3.8",
+    "async-timeout==5.0.1",
+    "asyncpg>=0.31.0",
+    "certifi>=2025.8.3",
+    "click==8.1.8",
+    "colorama==0.4.6",
+    "dill==0.3.9",
+    "exceptiongroup==1.2.2",
+    "fastapi==0.115.6",
+    "greenlet>=3.1.1",
+    "h11==0.16.0",
+    "httpcore==1.0.9",
+    "httpx==0.28.1",
+    "idna==3.10",
+    "iniconfig==2.1.0",
+    "isort==5.13.2",
+    "mako==1.3.8",
+    "markupsafe==3.0.2",
+    "mccabe==0.7.0",
+    "packaging==24.2",
+    "pathspec==0.12.1",
+    "platformdirs==4.3.6",
+    "pluggy==1.6.0",
+    "psycopg>=3.3.2",
+    "psycopg2-binary==2.9.10",
+    "pydantic==2.10.5",
+    "pydantic-core==2.27.2",
+    "pylint==3.3.3",
+    "pytest>=8.3.4",
+    "pytest-asyncio>=1.3.0",
+    "python-dotenv==1.0.1",
+    "sniffio==1.3.1",
+    "sqlalchemy[asyncio]>=2.0.37",
+    "starlette==0.41.3",
+    "tomli==2.2.1",
+    "tomlkit==0.13.2",
+    "typing-extensions==4.12.2",
+    "uvicorn[standard]==0.34.0",
+]
 
-[tool.poetry.dependencies]
-python = "^3.10"
-alembic = "1.14.0"
-annotated-types = "0.7.0"
-anyio = "4.8.0"
-astroid = "3.3.8"
-async-timeout = "5.0.1"
-asyncpg = "0.30.0"
-click = "8.1.8"
-dill = "0.3.9"
-exceptiongroup = "1.2.2"
-fastapi = "0.115.6"
-greenlet = "3.1.1"
-h11 = "0.14.0"
-idna = "3.10"
-isort = "5.13.2"
-mako = "1.3.8"
-markupsafe = "3.0.2"
-mccabe = "0.7.0"
-platformdirs = "4.3.6"
-psycopg2-binary = "2.9.10"
-pydantic = "2.10.5"
-pydantic-core = "2.27.2"
-pylint = "3.3.3"
-python-dotenv = "1.0.1"
-sniffio = "1.3.1"
-sqlalchemy = "2.0.37"
-starlette = "0.41.3"
-tomli = "2.2.1"
-tomlkit = "0.13.2"
-typing-extensions = "4.12.2"
-uvicorn = "0.34.0"
-
-
-[tool.poetry.group.dev.dependencies]
-black = "^25.1.0"
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
 
 [tool.black]
 line-length = 79
@@ -60,6 +62,40 @@ ignore_missing_imports = true
 
 [tool.pytest.ini_options]
 asyncio_mode = "auto"
-asyncio_default_fixture_loop_scope = "function"
+asyncio_default_fixture_loop_scope = "session"
 log_cli = 1
 log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
+addopts = [
+    "--strict-markers",
+    "--tb=short",
+    "-p", "no:warnings",
+    "--disable-warnings"
+]
+filterwarnings = [
+    "ignore::sqlalchemy.exc.SAWarning",
+    "ignore::RuntimeWarning"
+]
+
+[tool.ruff]
+line-length = 79
+indent-width = 4
+exclude = [
+    "alembic",
+    "migration",
+    "migrations"
+]
+
+
+[tool.ruff.lint]
+# Extend the default rule selection to include import sorting (I rules)
+extend-select = ["I"]
+fixable = ["I"]
+
+[tool.ruff.lint.pycodestyle]
+max-doc-length = 79
+
+[dependency-groups]
+dev = [
+    "isort>=5.13.2",
+    "ruff>=0.14.14",
+]
