@@ -1,5 +1,13 @@
-from sqlalchemy import Integer, String
-from sqlalchemy.orm import Mapped, declarative_base, mapped_column
+import datetime
+from typing import Optional
+
+from sqlalchemy import DateTime, ForeignKey, Integer, String, Text, func
+from sqlalchemy.orm import (
+    Mapped,
+    declarative_base,
+    mapped_column,
+    relationship,
+)
 
 Base = declarative_base()
 
@@ -15,3 +23,62 @@ class User(Base):
     active: Mapped[int] = mapped_column(
         Integer, nullable=False, server_default="1"
     )
+
+
+class Category(Base):
+    __tablename__ = "category"
+
+    id: Mapped[int] = mapped_column(primary_key=True)
+    title: Mapped[str] = mapped_column(Text(), nullable=False)
+    parent_id: Mapped[Optional[int]] = mapped_column(
+        ForeignKey("category.id", ondelete="SET NULL"),
+        nullable=True,
+        server_default="1",
+    )
+
+
+class Product(Base):
+    __tablename__ = "product"
+    id: Mapped[int] = mapped_column(primary_key=True)
+    title: Mapped[str] = mapped_column(String(50), nullable=False)
+    price: Mapped[str] = mapped_column(Integer(), nullable=False)
+    category_id: Mapped[int] = mapped_column(
+        ForeignKey("category.id", ondelete="RESTRICT"), nullable=False
+    )
+
+
+class Client(Base):
+    __tablename__ = "client"
+
+    id: Mapped[int] = mapped_column(primary_key=True)
+    name: Mapped[str] = mapped_column(String(40), nullable=False)
+    email: Mapped[str] = mapped_column(String(30), nullable=False)
+    address: Mapped[str] = mapped_column(String(120), nullable=True)
+
+
+class Order(Base):
+    __tablename__ = "order"
+    id: Mapped[int] = mapped_column(primary_key=True)
+    client_id: Mapped[int] = mapped_column(
+        Integer,
+        nullable=False,
+    )
+    date: Mapped[datetime.datetime] = mapped_column(
+        DateTime, nullable=False, server_default=func.now()
+    )
+
+
+class OrderItem(Base):
+    __tablename__ = "order_item"
+    id: Mapped[int] = mapped_column(primary_key=True)
+    order_id: Mapped[int] = mapped_column(
+        Integer, ForeignKey("order.id", ondelete="CASCADE"), nullable=False
+    )
+    product_id: Mapped[int] = mapped_column(
+        Integer, ForeignKey("product.id", ondelete="CASCADE"), nullable=False
+    )
+    quantity: Mapped[int] = mapped_column(Integer, server_default="1")
+    price_at_time: Mapped[int] = mapped_column(Integer, nullable=False)
+
+    # order = relationship("Order", back_populates="items")
+    # product = relationship("Product", back_populates="order_items")
