@@ -2,13 +2,10 @@ import uvicorn
 from fastapi import FastAPI
 
 from service.endpoints.data_handlers import api_router as data_routes
-from service.endpoints.put_handlers import api_router as put_routes
-from service.endpoints.update_handlers import api_router as upd_routes
 
 app = FastAPI()
 
-app.include_router(put_routes)
-app.include_router(upd_routes)
+
 app.include_router(data_routes)
 
 
